Create master branch at HEAD for repo being tested
For the repo which we are testing the build, we currently clone but do
not create a branch. This means that when building the build system
component `planex-pin update` fails, complaining that the master branch
does not exist.

To solve this, create a new master branch at the current HEAD.

Signed-off-by: John Else <john.else@citrix.com>